requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-extensions==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
parso==0.7.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptpython==3.0.2
Pygments==2.6.1
pyparsing==2.4.7
python-bitcoinrpc==1.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
